fix: correct the examples
