--- playground/src/pages/dynamic/$single.page.js
@@ -1,17 +1,0 @@
-import { html } from '@arrow-js/core'
-
-export const loader = ({ req }) => {
-  return {
-    single: req.params.single
-  }
-}
-
-export default function Page (loaderData) {
-  return html`
-    From single
-    <pre>
-      ${JSON.stringify(loaderData, null, 2)}
-    </pre
-    >
-  `
-}
